Refactor project and update dependencies
