+++ README.md
@@ -10,7 +10,7 @@ Configure `sample.env` and rename it to `.env`.
 $ virtualenv env
 $ source env/bin/activate
 $ pip install -r requirements.txt
-$ python main.py
+$ fastapi dev main:app
 ```
 
 - Docker compose:
