+++ requirements.txt
@@ -1,24 +1,21 @@
-annotated-types==0.7.0
-anyio==4.8.0
-click==8.1.8
-dnspython==2.7.0
-ecdsa==0.19.0
-exceptiongroup==1.2.2
-fastapi==0.115.8
-h11==0.14.0
-idna==3.10
-motor==3.7.0
-passlib==1.7.4
-pyasn1==0.6.1
-pydantic==2.10.6
-pydantic_core==2.27.2
-pymongo==4.11.1
-python-dotenv==1.0.1
+fastapi==0.111.0
+uvicorn==0.29.0
+pydantic==2.7.1
+pydantic_core==2.16.3
+motor==3.4.0
 python-jose==3.3.0
-python-multipart==0.0.20
+passlib==1.7.4
+python-multipart==0.0.9
+python-dotenv==1.0.1
+starlette==0.37.2
+anyio==4.3.0
+click==8.1.7
+dnspython==2.6.1
+ecdsa==0.18.0
+h11==0.14.0
+idna==3.7
+pymongo==4.5.0
 rsa==4.9
 six==1.17.0
 sniffio==1.3.1
-starlette==0.45.3
 typing_extensions==4.12.2
-uvicorn==0.34.0
